requirements.txt
asgi-redis==1.0.0
asgiref==3.4.1
asn1crypto==0.24.0
astroid==2.2.5
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
Brlapi==0.6.6
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2018.1.18
cffi==1.15.0
channels==3.0.4
chardet==3.0.4
constantly==15.1.0
cryptography==36.0.0
cssselect2==0.4.1
cupshelpers==1.0
daphne==3.0.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-extensions==0.7.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.9.2
hyperlink==21.0.0
idna==2.6
image==1.5.33
importlib-metadata==4.6.4
incremental==21.3.0
isort==4.3.17
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
msgpack-python==0.5.6
netifaces==0.10.4
notify2==0.3
oauth==1.0.1
olefile==0.45.1
pbr==5.6.0
pexpect==4.2.1
Pillow==8.4.0
protobuf==3.0.0
psycopg2==2.7.7
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
PyJWT==2.3.0
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==21.0.0
pyphen==0.11.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-debian==0.1.32
python-decouple==3.4
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
redis==4.0.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.13.2
six==1.11.0
sqlparse==0.4.1
stevedore==3.4.0
system-service==0.3
systemd-python==234
testresources==2.0.1
tinycss2==1.1.0
twilio==7.2.0
Twisted==21.7.0
txaio==21.2.1
typed-ast==1.3.4
typing-extensions==3.10.0.0
ubuntu-advantage-tools==27.2
ufw==0.36
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
virtualenv-clone==0.5.6
virtualenvwrapper==4.8.4
wadllib==1.3.2
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.11.1
xkit==0.0.0
youtube_dl==2018.3.14
zipp==3.5.0
zope.interface==5.4.0
